move govulncheck to reachability
