+++ enricher/reachability/go/source/gvcinterface.go
@@ -17,7 +17,7 @@ package source
 import (
 	"context"
 
-	"github.com/google/osv-scalibr/enricher/govulncheck/source/internal"
+	"github.com/google/osv-scalibr/enricher/reachability/go/source/internal"
 	vulnpb "github.com/ossf/osv-schema/bindings/go/osvschema"
 )
