+++ enricher/reachability/go/source/gvcreal.go
@@ -21,8 +21,8 @@ import (
 	"os"
 	"os/exec"
 
-	"github.com/google/osv-scalibr/enricher/govulncheck/source/internal"
-	"github.com/google/osv-scalibr/enricher/govulncheck/source/internal/url"
+	"github.com/google/osv-scalibr/enricher/reachability/go/source/internal"
+	"github.com/google/osv-scalibr/enricher/reachability/go/source/internal/url"
 	vulnpb "github.com/ossf/osv-schema/bindings/go/osvschema"
 	"golang.org/x/vuln/scan"
 	"google.golang.org/protobuf/encoding/protojson"
