+++ enricher/reachability/go/source/internal/result.go
