+++ enricher/reachability/go/source/testdata/govulncheckinput.json
